Flushing a filter does not acquire a read lock
Old behavior was to increment the reference count to a filter
and acquire a read lock. The unintended side effect of this is
that writes end up bounded by Disk I/O speed, as the lock is not
released until the flush completes. To fix this, we increment
the ref count, and thus make sure the filter is not deleted,
but do not acquire a read or write lock. This allows operations to
continue normally, bounded by memory instead of disk.