Fixing send_client_response when num_bufs > IOV_MAX
Previously, there was no limit to the number of buffers that
would try to get passed to writev(), however the kernel
does export IOV_MAX as the limit. We did not notice this limitation
because older builds of glibc works around by copying to a large buffer.
On Ubuntu 12.04 this behavior changed and broke, so now we limit
the calls to writev().