First pass at adding MVCC for filter views
Previously, we had a single shared hashmap protected with
a spin lock, and each filter used reference counts. In an attempt
to increase concurrency, we are removing this and adding a
form of MVCC. Instead, each hash map has a version number,
and we can allow multiple versions to exist at a time. This
should reduce the locking required for reads. Particularly,
this will be a massive win for doing lists without causing
massive read contention for the spin lock. Writes must increment
the version, and perform a pointer swap. This does have the down
side of forcing copies of the entire hash map on changes, however
the relative infrequency of this offsets the reduced locking.