Removes is_hot on versions, revert 78e679af921ad481019122516876967d509fc947
Basically a revert of 78e679af921ad481019122516876967d509fc947. This
heuristic is actually very unsafe and has very poor characteristics.
Basically we are relying on operations to not take more than an
aribtrary cool down period and that sleep is predictable. Both of these
are patently false, and in retrospect, I have no idea what I was
thinking. This mechanism needs to be replaced with something much more
reliable and determininistic.