Remove leader/follower, used fixed assignment
We are ripping out the old leader/follower model and replacing it
with a fixed connection thread assignment. The old model was designed
to avoid a single slow operaiton from blocking all the handlers, but
likely if we are hitting a slow operation, all the threads will be
exhausted regardless. The old model had a high complexity overhead
to maintain thread safety in the case of multiple threads accessing
buffers and connections for a given client. The new model assigns each
client to a specific thread, each of which have their own libev
event loop. This simplifies the code, reduces locking and should
improve performance under high contention.