Fix #5 - Style nit, 'const' for int param
For some reason, MurmurHash3_x64_128() has 'const' in front of 'seed',
which is an integer type. This doesn't seem like it will yield any
compiler optimizations at all, especially given that we are NOT inlining
the MurmurHash functions. Additionally, this doesn't match any of the
other hash functions, nor the header's prototype of the function. So
just drop the 'const'.

CR: gcc, tests