New interface to register filt manager clients
Adding a new interface that uses `filtmgr_client_checkpoint`
and `filtmgr_client_leave`. Checkpoint is used to register and
periodically update the latest seen version, while leave is used
when a thread no long needs to make use of the filter manager.

This change allows the vacuum thread to more safely determine
the last version of all clients. Previously, the client list was
limited to the workers using a fixed size array. This approach
did not allow the flush and cold unmap threads to be accounted for,
and thus had potential safely problems. The new interface is more
accomadating and can even allow for things like temporary threads
that need access to the filter manager.