Better handling of an in-progress delete
We now do a much better job of detecting that a delete is in
progress and avoiding doing a conflicting create at the same time.
Previously, due to the MVCC architecture it was possible to have
a filter be deleted in version V, and then created in V+1, even
though the vacuum thread has not yet cleared version V.

This would cause a duplication of the internal buffers, and eventually
the vacuum thread will do a delete of version V, and the on-disk
contents would be lost.